requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.2.0
backoff==1.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.9
frozenlist==1.3.3
fuzzywuzzy==0.18.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.21.1
MarkupSafe==2.1.1
multidict==6.0.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-Levenshtein==0.21.1
pytz==2022.7
rapidfuzz==3.2.0
requests==2.31.0
six==1.16.0
spotify==0.10.2
tidalapi==0.7.2
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.8.2
